Modify double type migration to float for PG in Heroku

+++ db/migrate/20150519135320_add_weight_to_exercise_sets.rb
@@ -1,5 +1,5 @@
 class AddWeightToExerciseSets < ActiveRecord::Migration
   def change
-    add_column :exercise_sets, :weight, :double
+    add_column :exercise_sets, :weight, :float
   end
 end
